perf: 修改部门接口
1.引入DeptConverter实现部门实体与视图对象的转换

2. 优化DeptService.getDeptTree方法，增强代码可读性与性能

3. 修改部门级别和父部门等级的数据类型为Short

4. 使用常量替代硬编码的数字，增强代码的可读性和维护性

5.修复部门名称重复校验的SQL查询条件错误

+++ src/main/java/com/example/workflow/model/dto/DeptFormDto.java
@@ -13,7 +13,7 @@ public class DeptFormDto {
     /**
      * 父部门等级
      */
-    private Integer parentLevel;
+    private Short parentLevel;
     /**
      * 部门名称
      */
